Separate newAdmin function and command-line runner
